requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.18
django-bootstrap3==23.1
django-bootstrap5==22.2
django-countries==7.5.1
django-filter==22.1
django-tables2==2.5.2
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
typing-extensions==4.5.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
